+++ ams/routines/ed.py
@@ -46,9 +46,9 @@ class SRBase:
 
         # NOTE: define e_str in dispatch model
         self.prsb = Constraint(info='spinning reserve balance',
-                               name='prsb', type='eq',)
+                               name='prsb', is_eq=True,)
         self.rsr = Constraint(info='spinning reserve requirement',
-                              name='rsr', type='uq',)
+                              name='rsr', is_eq=False,)
 
 
 class MPBase:
@@ -109,7 +109,7 @@ class MPBase:
         self.aBus.info = '2D Bus angle'
 
 
-class ED(RTED):
+class ED(RTED, MPBase, SRBase):
     """
     DC-based multi-period economic dispatch (ED).
     Dispath interval ``config.t`` (:math:`T_{cfg}`) is introduced,
@@ -202,11 +202,11 @@ class ED(RTED):
         self.rgu0 = Constraint(name='rgu0',
                                info='Initial gen ramping up',
                                e_str='pg[:, 0] - pg0[:, 0] - R30',
-                               type='uq',)
+                               is_eq=False,)
         self.rgd0 = Constraint(name='rgd0',
                                info='Initial gen ramping down',
                                e_str='- pg[:, 0] + pg0[:, 0] - R30',
-                               type='uq',)
+                               is_eq=False,)
 
         # --- objective ---
         cost = 'sum(t**2 dot c2 @ pg**2)'
@@ -279,11 +279,11 @@ class ESD1MPBase(ESD1Base):
 
         SOCb0 = 'mul(En, SOC[:, 0] - SOCinit) - t dot mul(EtaC, zce[:, 0])'
         SOCb0 += ' + t dot mul(REtaD, zde[:, 0])'
-        self.SOCb0 = Constraint(name='SOCb0', type='eq',
+        self.SOCb0 = Constraint(name='SOCb0', is_eq=True,
                                 info='ESD1 SOC initial balance',
                                 e_str=SOCb0,)
 
-        self.SOCr = Constraint(name='SOCr', type='eq',
+        self.SOCr = Constraint(name='SOCr', is_eq=True,
                                info='SOC requirement',
                                e_str='SOC[:, -1] - SOCinit',)